Fix: groupId / OrderId INT로 수정

+++ src/main/java/com/somoa/serviceback/domain/order/dto/OrderSaveDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class OrderSaveDto {
 
-    private Long groupId;
+    private Integer groupId;
     private Integer supplyId;
     private String orderStatus;
     private String productName;
